Add files via upload
This folder contains the data model used in the SSIS project for the Global Superstore dataset. The image, shows the relationships between the key tables in the database: FactOrders, which holds the transaction details, and the dimension tables (DimCustomers, DimProducts, DimTime, DimLocations) that provide context to the facts. The schema is designed to ensure data is well-organized and easy to analyze, making it a crucial part of the ETL process and the overall data warehouse structure. This visual representation helps everyone involved in the project to understand how the data is structured and how different pieces fit together.